Fixed more bugs. Still at least one more remaining.
